refactor: 신체 정보 입력 화면 클래스 및 파일명 변경
