+++ android/app/src/main/java/com/mulkkam/ui/onboarding/OnboardingStep.kt
@@ -1,8 +1,8 @@
 package com.mulkkam.ui.onboarding
 
 import androidx.fragment.app.Fragment
+import com.mulkkam.ui.onboarding.bioinfo.BioInfoFragment
 import com.mulkkam.ui.onboarding.nickname.NicknameFragment
-import com.mulkkam.ui.onboarding.physicalinfo.PhysicalInfoFragment
 import com.mulkkam.ui.onboarding.targetamount.TargetAmountFragment
 import com.mulkkam.ui.onboarding.terms.TermsFragment
 
@@ -11,6 +11,6 @@ enum class OnboardingStep(
 ) {
     TERMS({ TermsFragment() }),
     NICKNAME({ NicknameFragment() }),
-    PHYSICAL_INFO({ PhysicalInfoFragment() }),
+    PHYSICAL_INFO({ BioInfoFragment() }),
     TARGET_AMOUNT({ TargetAmountFragment() }),
 }
